pyproject.toml
[project]
name = "SoraCleaner"
version = "0.1.0"
description = "A watermark removal tool specifically designed for Sora2. "
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "basicsr>=1.4.2",
    "matplotlib>=3.10.6",
    "torch",
    "opencv-python",
    "numpy",
    "Pillow",
    "PyYAML",
    "tqdm",
    "av",
]
